% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcAllometries.R
\name{allometries}
\alias{allometries}
\title{lfcAllometries class}
\usage{
allometries()
}
\value{
An \code{lfcAllometries} class object (inherits from \code{\link[R6]{R6Class}}),
  with methods to access the data. See Methods section.
}
\description{
\code{allometries()} creates an object to access the allometries database.
}
\section{Methods}{

  \code{lfcAllometries} objects has two public methods:
  \itemize{
    \item{\code{$get_data}: Retrieve and collect the allometries table. See
          \code{\link{allometries_get_data}} for more details}
    \item{\code{$description}: Retrieve the description for the desired
          allometries. See \code{\link{allometries_description}} for
          more details}
    \item{\code{$calculate}: Calculate variables based on the selected
          allometries. See \code{\link{allometries_calculate}} for
          more details}
    \item{\code{$describe_var}: Print the information available about the provided
          variable. See \code{\link{allometries_describe_var}} for more details}
  }
}

\examples{
allomdb <- allometries()
allomdb
}
\seealso{
Other Allometries functions: 
\code{\link{allometries_calculate}()},
\code{\link{allometries_description}()},
\code{\link{allometries_get_data}()}
}
\concept{Allometries functions}
