% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcAllometries.R
\name{allometries_calculate}
\alias{allometries_calculate}
\title{Calculating new variables based on the allometries formula}
\usage{
allometries_calculate(object, ..., allometry_id)
}
\arguments{
\item{object}{\code{lfcAllometries} object, as created by \code{\link{allometries}}}

\item{...}{\bold{Must be named}. Numeric vectors for the independent variables
present in the allometry equation. Each argument must be named as the
independent variable the values correspond with. See examples.}

\item{allometry_id}{character with the unique allometry identifier}
}
\description{
Return a vector with the desired allometry equation calculated
}
\examples{

if (interactive()) {
  library(dplyr)
  allomdb <- allometries()
  allometries_calculate(allomdb, DR = c(0.55, 0.46, 0.37), allometry_id = "BH_287")

  # inside a dplyr mutate, with a different allometry for each species
  iris_foo <- iris \%>\%
    mutate(allom = rep(c("BH_287","BH_288","BH_290"), each = 50)) \%>\%
    select(branch_diameter = Sepal.Length, Species, allom)
  iris_foo

  iris_foo \%>\%
    group_by(Species) \%>\%
    mutate(BH = allometries_calculate(
        allomdb, DR = branch_diameter, allometry_id = first(allom)
    ))
}

}
\seealso{
Other Allometries functions: 
\code{\link{allometries_description}()},
\code{\link{allometries_get_data}()},
\code{\link{allometries}()}
}
\concept{Allometries functions}
