% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcAllometries.R
\name{allometries_describe_var}
\alias{allometries_describe_var}
\title{Print info about the variables present in the allometries db}
\usage{
allometries_describe_var(object, variables)
}
\arguments{
\item{object}{\code{lfcAllometries} object, as created by \code{\link{allometries}}}

\item{variables}{character vector with the names of the variables to describe}
}
\value{
A character vector with the variable names to describe
}
\description{
\code{allometries_describe_var} is a wrapper for the \code{$describe_var} method of
  \code{lfcAllometries} objects. See \code{\link{allometries}}.
}
\examples{
if (interactive()) {
allometriesdb <- allometries()
allometries_describe_var(allometriesdb, "BR")
allometries_describe_var(allometriesdb, c("DBH", "P_BST"))

# allometriesdb is an R6 object, so the previous example is the same as:
allometriesdb$describe_var("BR")
allometriesdb$describe_var(c("DBH", "P_BST"))
}

}
\concept{allometries functions}
