% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcAllometries.R
\name{allometries_description}
\alias{allometries_description}
\title{Info for allometries}
\usage{
allometries_description(object, ..., id = NULL)
}
\arguments{
\item{object}{\code{lfcAllometries} object, as created by \code{\link{allometries}}}

\item{...}{Expresions resolving to a logical value to filter the allometries. Only
evaluated if \code{id} is NULL.}

\item{id}{Character vector with the allometry/ies id}
}
\value{
A list with the selected allometries and their info, see examples. In the
  special case of no \code{id} or expressions (\code{...}) provided, information for
  all the allometries is returned.
}
\description{
\code{allometries_description} is a wrapper for the \code{$description}
  method of \code{lfcAllometries} objects. See also \code{\link{allometries}}.
}
\examples{

if (interactive()) {
  # by id
  allomdb <- allometries()
  foo <- allometries_description(allomdb, id = "GC_3210")
  foo
  foo$GC_3210$dependent_var
  foo$GC_3210$param_a

  # filtering
  ht_dn_allometries <- allometries_description(allomdb, dependent_var \%in\% c("GC", "Dn"))
  length(ht_dn_allometries)
  names(ht_dn_allometries)
  ht_dn_allometries$GC_3210$dependent_var
  ht_dn_allometries$GC_3210$param_a
}

}
\seealso{
Other Allometries functions: 
\code{\link{allometries_calculate}()},
\code{\link{allometries_get_data}()},
\code{\link{allometries}()}
}
\concept{Allometries functions}
