% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcAllometries.R
\name{allometries_get_data}
\alias{allometries_get_data}
\title{Access to the tables in the allometries database}
\usage{
allometries_get_data(object, table_name = "allometries")
}
\arguments{
\item{object}{\code{lfcAllometries} object, as created by \code{\link{allometries}}}

\item{table_name}{character vector of lenght 1 indicating the requested table name.
Only value available at the moment is \code{"allometries"}}
}
\value{
A tbl object
}
\description{
\code{allometries_get_data} is a wrapper for the \code{$get_data} method of
  \code{lfcAllometries} objects. See also \code{\link{allometries}}.
}
\details{
Connection to database can be slow. Tables retrieved from the db are stored
  in a temporary cache inside the lfcAllometries object created by
  \code{\link{allometries}}, making subsequent calls to the same table are faster.
}
\examples{
if (interactive()) {
  allomdb <- allometries()
  # tibble
  allometries_get_data(allomdb, 'allometries')

  # we can use pipes
  allomdb \%>\%
    allometries_get_data('allometries')

  # allomdb is an R6 object, so the previous examples are the same as:
  allomdb$get_data('allometries')
}

}
\seealso{
Other Allometries functions: 
\code{\link{allometries_calculate}()},
\code{\link{allometries_description}()},
\code{\link{allometries}()}
}
\concept{Allometries functions}
