% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcCatDrought.R
\name{catdrought}
\alias{catdrought}
\title{lfcCatDrought class}
\usage{
catdrought()
}
\value{
An \code{lfcCatDrought} class object (inherits from
  \code{\link[R6]{R6Class}}), with methods to access the data. See Methods
  section.
}
\description{
\code{catdrought()} creates an object to access the CatDrought
  database.
}
\section{Methods}{

  \code{lfcCatDrought} objects have the following methods available:
  \itemize{
    \item{\code{$get_data}: Returns none, maintaned for consistency}
    \item{\code{$get_raster}: Returns the raster for the selected date}
    \item{\code{$get_current_time_series}: Returns a dataframe with the
    time series for a provided spatial object (points or polygons)}
  }
}

\examples{
catdroughtdb <- catdrought()
catdroughtdb
}
\seealso{
Other catdrought functions: 
\code{\link{catdrought_describe_var}()},
\code{\link{catdrought_get_current_time_series}()},
\code{\link{catdrought_get_raster}()}
}
\concept{catdrought functions}
