% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcCatDrought.R
\name{catdrought_describe_var}
\alias{catdrought_describe_var}
\title{Print info about the variables present in the CatDrought db}
\usage{
catdrought_describe_var(object, variables)
}
\arguments{
\item{object}{\code{lfcCatDrought} object, as created by \code{\link{catdrought}}}

\item{variables}{character vector with the names of the variables to describe}
}
\value{
A character vector with the variable names to describe
}
\description{
\code{catdrought_describe_var} is a wrapper for the \code{$describe_var} method
  of \code{lfcCatDrought} objects. See \code{\link{catdrought}}.
}
\examples{
if (interactive()) {
catdroughtdb <- catdrought()
catdrought_describe_var(catdroughtdb, "DDS")
catdrought_describe_var(catdroughtdb, c("Esoil", "REW"))

# catdroughtdb is an R6 object, so the previous example is the same as:
catdroughtdb$describe_var("DDS")
catdroughtdb$describe_var(c("Esoil", "REW"))
}

}
\seealso{
Other catdrought functions: 
\code{\link{catdrought_get_current_time_series}()},
\code{\link{catdrought_get_raster}()},
\code{\link{catdrought}()}
}
\concept{catdrought functions}
