% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcCatDrought.R
\name{catdrought_get_current_time_series}
\alias{catdrought_get_current_time_series}
\title{Create time series for CatDrought variables for the current year}
\usage{
catdrought_get_current_time_series(object, sf, variable)
}
\arguments{
\item{object}{\code{lfcCatDrought} object, as created by
\code{\link{catdrought}}}

\item{sf}{sf object with polygons or points where to calculate the time
series}

\item{variable}{character indicating the desired variable to create the
time series. It should be one of 'DDS', 'DeepDrainage', 'Eplant', 'Esoil',
'Infiltration', 'LAI', 'PET', 'Psi', 'REW', 'Runoff' or 'Theta'}
}
\value{
A data frame with the date, sf identification and the
variable value for points or the mean and standard error values for polygons.
}
\description{
\code{catdrought_get_time_series} is a wrapper for the
  \code{$get_current_time_series} method of \code{lfcCatDrought} objects.
  See also \code{\link{catdrought}}.
}
\details{
Calculations can be long depending on the number of features and/or
  size of polygons.
}
\section{sf}{

  sf objects must have a column with unique values for each feature as an
  identifier. This must be the first column in the sf object.
}

\examples{
if (interactive()) {
  # TODO examples
}

}
\seealso{
Other catdrought functions: 
\code{\link{catdrought_describe_var}()},
\code{\link{catdrought_get_raster}()},
\code{\link{catdrought}()}
}
\concept{catdrought functions}
