% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcFES.R
\name{fes}
\alias{fes}
\title{lfcFES class}
\usage{
fes()
}
\value{
An \code{lfcFES} class object (inherits from
  \code{\link[R6]{R6Class}}), with methods to access the data. See Methods
  section.
}
\description{
\code{fes()} creates an object to access the fes database.
}
\section{Methods}{

  \code{lfcFES} objects has the following public methods:
  \itemize{
    \item{\code{$get_data}: Retrieve and collect FES database tables. See
          \code{\link{fes_get_data}} for more details}
    \item{\code{$avail_tables}: Return a character vector with the names of
          the available tables in the database. See
          \code{\link{fes_avail_tables}} for more details}
    \item{\code{$describe_table}: Print the information available about the
          provided table. See \code{\link{fes_describe_table}} for more
          details}
    \item{\code{$describe_var}: Print the information available about the
          provided variable. See \code{\link{fes_describe_var}} for more
          details}
  }
}

\examples{
fesdb <- fes()
fesdb
}
\seealso{
Other FES functions: 
\code{\link{fes_avail_tables}()},
\code{\link{fes_describe_var}()},
\code{\link{fes_get_data}()}
}
\concept{FES functions}
