% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcFES.R
\name{fes_describe_table}
\alias{fes_describe_table}
\title{Print info about the tables present in the FES db}
\usage{
fes_describe_table(object, tables)
}
\arguments{
\item{object}{\code{lfcFES} object, as created by \code{\link{fes}}}

\item{tables}{character vector with the names of the tables to describe}
}
\value{
Description is printed in the console, nothing is returned
}
\description{
\code{fes_describe_table} is a wrapper for the \code{$describe_table}
  method of \code{lfcFES} objects. See \code{\link{fes}}.
}
\examples{
if (interactive()) {
  fesdb <- fes()
  fes_describe_table(fesdb, "static")
  fes_describe_table(fesdb, c("static", "plot_nfi_2_results"))

  # fesdb is an R6 object, so the previous example is the same as:
  fesdb$describe_table("static")
  fesdb$describe_table(c("static", "plot_nfi_2_results"))
}

}
\concept{fes functions}
