% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcFES.R
\name{fes_describe_var}
\alias{fes_describe_var}
\title{Print info about the variables present in the FES db}
\usage{
fes_describe_var(object, variables)
}
\arguments{
\item{object}{\code{lfcFES} object, as created by \code{\link{fes}}}

\item{variables}{character vector with the names of the variables to describe}
}
\value{
Description is printed in the console, nothing is returned
}
\description{
\code{fes_describe_var} is a wrapper for the \code{$describe_var}
  method of \code{lfcFES} objects. See \code{\link{fes}}.
}
\examples{
if (interactive()) {
  fesdb <- fes()
  fes_describe_var(fesdb, "mushrooms_poduction")
  fes_describe_var(fesdb, c("exported_water", "animals_presence"))

  # fesdb is an R6 object, so the previous example is the same as:
  fesdb$describe_var("mushrooms_poduction")
  fesdb$describe_var(c("exported_water", "animals_presence"))
}

}
\seealso{
Other FES functions: 
\code{\link{fes_avail_tables}()},
\code{\link{fes_get_data}()},
\code{\link{fes}()}
}
\concept{FES functions}
