% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcFES.R
\name{fes_get_data}
\alias{fes_get_data}
\title{Access to the tables in the FES database}
\usage{
fes_get_data(object, table_name, spatial = FALSE)
}
\arguments{
\item{object}{\code{lfcFES} object, as created by \code{\link{fes}}}

\item{table_name}{character vector of lenght 1 indicating the requested table
name}

\item{spatial}{logical indicating if the data must be converted to an spatial
object}
}
\value{
A tbl object: tbl_df if spatial is \code{FALSE}, sf if spatial is
  \code{TRUE}
}
\description{
\code{fes_get_data} is a wrapper for the \code{$get_data} method
  of \code{lfcFES} objects. See also \code{\link{fes}}.
}
\details{
Connection to database can be slow. Tables retrieved from the db are
  stored in a temporary cache inside the lfcFES object created by
  \code{\link{fes}}, making subsequent calls to the same table are faster.
}
\examples{
if (interactive()) {
  fesdb <- fes()
  # tibble
  fes_get_data(fesdb, 'static')
  # sf tibble
  fes_get_data(fesdb, 'static', TRUE)

  # we can use pipes
  fesdb \%>\%
    fes_get_data('static', TRUE)

  # fesdb is an R6 object, so the previous examples are the same as:
  fesdb$get_data('static')
  fesdb$get_data('static', TRUE)
}

}
\seealso{
Other FES functions: 
\code{\link{fes_avail_tables}()},
\code{\link{fes_describe_var}()},
\code{\link{fes}()}
}
\concept{FES functions}
