% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcLiDAR.R
\name{lidar_describe_var}
\alias{lidar_describe_var}
\title{Print info about the variables present in the LiDAR db}
\usage{
lidar_describe_var(object, variables)
}
\arguments{
\item{object}{\code{lfcLiDAR} object, as created by \code{\link{lidar}}}

\item{variables}{character vector with the names of the variables to describe}
}
\value{
A character vector with the variable names to describe
}
\description{
\code{lidar_describe_var} is a wrapper for the \code{$describe_var} method
  of \code{lfcLiDAR} objects. See \code{\link{lidar}}.
}
\examples{
if (interactive()) {
lidardb <- lidar()
lidar_describe_var(lidardb, "BF")
lidar_describe_var(lidardb, c("DBH", "VAE"))

# lidardb is an R6 object, so the previous example is the same as:
lidardb$describe_var("BF")
lidardb$describe_var(c("DBH", "VAE"))
}

}
\seealso{
Other LiDAR functions: 
\code{\link{lidar_avail_tables}()},
\code{\link{lidar_clip_and_stats}()},
\code{\link{lidar_get_data}()},
\code{\link{lidar_get_lowres_raster}()},
\code{\link{lidar_point_value}()},
\code{\link{lidar}()}
}
\concept{LiDAR functions}
