% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcLiDAR.R
\name{lidar_get_data}
\alias{lidar_get_data}
\title{Access the aggregated data for administrative and natural areas}
\usage{
lidar_get_data(
  object,
  table_name,
  variables = c("AB", "BAT", "BF", "CAT", "DBH", "HM", "REC", "VAE")
)
}
\arguments{
\item{object}{\code{lfcLiDAR} object, as created by \code{\link{lidar}}}

\item{table_name}{character vector of lenght 1 indicating the table to retrieve}

\item{variables}{character vector indicating variables for which data is returned. If
not provided, all variables stats are returned}
}
\value{
An sf object with the aggregated values for each administrative division or
  natural area for the variables requested
}
\description{
\code{lidar_get_data} is a wrapper for the \code{$get_data} method of
  \code{lfcLiDAR} objects. See also \code{\link{lidar}}.
}
\details{
Precalculated aggregated values for
  \itemize{
    \item{Catalonia, in the \code{lidar_catalunya} table}
    \item{Provinces, in the \code{lidar_provincias} table}
    \item{Veguerias, in the \code{lidar_veguerias} table}
    \item{Regions, in the \code{lidar_comarcas} table}
    \item{Municipalities, in the \code{lidar_municipalities} table}
    \item{National Parks}
    \item{Natura 2000 Network}
  }
}
\examples{
if (interactive()) {
  lidardb <- lidar()
  # provinces data for DBH and AB
  provinces_data <- lidar_get_data(lidardb, 'lidar_provincias', c('AB', 'DBH'))
  provinces_data

  # lidardb is an R6 object, so the previous example is the same as:
  lidardb$get_data('lidar_provincias', c('AB', 'DBH'))
}

}
\seealso{
Other LiDAR functions: 
\code{\link{lidar_avail_tables}()},
\code{\link{lidar_clip_and_stats}()},
\code{\link{lidar_describe_var}()},
\code{\link{lidar_get_lowres_raster}()},
\code{\link{lidar_point_value}()},
\code{\link{lidar}()}
}
\concept{LiDAR functions}
