% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcLiDAR.R
\name{lidar_get_lowres_raster}
\alias{lidar_get_lowres_raster}
\title{Access to the low resolution (400x400m) rasters in the LiDAR database}
\usage{
lidar_get_lowres_raster(object, variables, spatial = "stars")
}
\arguments{
\item{object}{\code{lfcLiDAR} object, as created by \code{\link{lidar}}}

\item{variables}{character vector indicating the requested raster/s variables.
\code{"all"} can be used to retrieve all variables.}

\item{spatial}{character vector of lenght 1 indicating the type of raster object to
return, "raster" or "stars", the default.}
}
\value{
A raster object: \code{RasterLayer} if spatial is \code{raster} and only one
  variable is requested, \code{RasterBrick} if more than one variable is requested.
  \code{stars} if spatial is \code{stars}. See
  https://r-spatial.github.io/stars/index.html for details about stars objects and
  \code{\link[raster]{raster}} for details about raster objects.
}
\description{
\code{lidar_get_lowres_raster} is a wrapper for the
  \code{$get_lowres_raster} method of \code{lfcLiDAR} objects.
  See also \code{\link{lidar}}.
}
\details{
Connection to database can be slow. Rasters retrieved from the db are stored
  in a temporary cache inside the lfcLiDAR object created by \code{\link{lidar}},
  making subsequent calls to the same table are faster. But, be warned that in-memory
  rasters can use a lot of memory!
}
\examples{
if (interactive()) {
  lidardb <- lidar()
  # raster
  ab_raster <- lidar_get_lowres_raster(lidardb, 'AB', 'raster')
  # stars
  ab_stars <- lidar_get_lowres_raster(lidardb, 'AB', 'stars')

  # we can use pipes
  lidardb \%>\%
    lidar_get_lowres_raster('AB', 'raster')

  # or retrieve several tables at one time
  lidardb \%>\%
    lidar_get_lowres_raster(c('AB', 'DBH'), 'stars')

  # lidardb is an R6 object, so the previous examples are the same as:
  lidardb$get_lowres_raster('AB', 'raster')
  lidardb$get_lowres_raster('AB', 'stars')
}

}
\seealso{
Other LiDAR functions: 
\code{\link{lidar_avail_tables}()},
\code{\link{lidar_clip_and_stats}()},
\code{\link{lidar_describe_var}()},
\code{\link{lidar_get_data}()},
\code{\link{lidar_point_value}()},
\code{\link{lidar}()}
}
\concept{LiDAR functions}
