% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcLiDAR.R
\name{lidar_point_value}
\alias{lidar_point_value}
\title{Extract point value from LiDAR raw raster tables (20x20m)}
\usage{
lidar_point_value(object, sf, point_id_variable, variables)
}
\arguments{
\item{object}{\code{lfcLiDAR} object, as created by \code{\link{lidar}}}

\item{sf}{sf object with the points to extract}

\item{point_id_variable}{character vector of length 1 with the name of the
variable of \code{sf} that contains the point identificator (name, code...)}

\item{variables}{character vector with the names of the variables to access}
}
\value{
This function returns the same sf object provided, with new columns with the
  values of each point for each variable requested.
}
\description{
\code{lidar_point_value} is a wrapper for the \code{$point_value}
  method of \code{lfcLiDAR} objects. See \code{\link{lidar}}.
}
\examples{
if (interactive()) {
library(dplyr)
lidardb <- lidar()

points_data <-
  nfi()$get_data('plots', spatial = TRUE) \%>\%
  dplyr::slice(1:5) \%>\%
  dplyr::select(plot_id)

dbh_plots <- lidar_point_value(lidardb, points_data, 'plot_id', 'DBH')

# lidardb is an R6 object, so the previous example is the same as:
lidardb$point_value(points_data, 'plot_id', 'DBH')
}

}
\seealso{
Other LiDAR functions: 
\code{\link{lidar_avail_tables}()},
\code{\link{lidar_clip_and_stats}()},
\code{\link{lidar_describe_var}()},
\code{\link{lidar_get_data}()},
\code{\link{lidar_get_lowres_raster}()},
\code{\link{lidar}()}
}
\concept{LiDAR functions}
