% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcMeteoland.R
\name{meteoland}
\alias{meteoland}
\title{lfcMeteoland class}
\usage{
meteoland()
}
\value{
An \code{lfcMeteoland} class object (inherits from
  \code{\link[R6]{R6Class}}), with methods to access the data. See Methods
  section.
}
\description{
\code{meteoland()} creates an object to access the Meteoland
  database.
}
\section{Methods}{

  \code{lfcMeteoland} objects have the following methods available:
  \itemize{
    \item{\code{$get_data}: }
  }
}

\examples{
meteolanddb <- meteoland()
meteolanddb
}
\concept{meteoland functions}
