% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcMeteoland.R
\name{meteoland_get_lowres_raster}
\alias{meteoland_get_lowres_raster}
\title{Access to the low resolution (1000x1000m) rasters in the Meteoland database}
\usage{
meteoland_get_lowres_raster(object, date, spatial = "stars")
}
\arguments{
\item{object}{\code{lfcMeteoland} object, as created by
\code{\link{meteoland}}}

\item{date}{character with the date of the raster to retrieve, i.e "2020-04-25"}

\item{spatial}{character vector of lenght 1 indicating the type of raster
object to return, "raster" or "stars", the default.}
}
\value{
A raster object: \code{RasterBrick} if spatial is \code{raster},
  \code{stars} if spatial is \code{stars}. See
  https://r-spatial.github.io/stars/index.html for details about stars
  objects and \code{\link[raster]{raster}} for details about raster objects.
}
\description{
\code{meteoland_get_lowres_raster} is a wrapper for the
  \code{$get_lowres_raster} method of \code{lfcMeteoland} objects.
  See also \code{\link{meteoland}}.
}
\details{
Connection to database can be slow. Rasters retrieved from the db
  are stored in a temporary cache inside the lfcMeteoland object created by
  \code{\link{meteoland}}, making subsequent calls to the same table are
  faster. But, be warned that in-memory rasters can use a lot of memory!
}
\examples{
if (interactive()) {
  meteolanddb <- meteoland()
  # raster
  ab_raster <- meteoland_get_lowres_raster(meteolanddb, '2020-04-25', 'raster')
  # stars
  ab_stars <- meteoland_get_lowres_raster(meteolanddb, '2020-04-25', 'stars')

  # we can use pipes
  meteolanddb \%>\%
    meteoland_get_lowres_raster('2020-04-25', 'raster')

  # meteolanddb is an R6 object, so the previous examples are the same as:
  meteolanddb$get_lowres_raster('2020-04-25', 'raster')
  meteolanddb$get_lowres_raster('2020-04-25', 'stars')
}

}
\seealso{
Other Meteoland functions: 
\code{\link{meteoland_historical_points_interpolation}()},
\code{\link{meteoland_points_interpolation}()},
\code{\link{meteoland_raster_interpolation}()}
}
\concept{Meteoland functions}
