% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcMeteoland.R
\name{meteoland_historical_points_interpolation}
\alias{meteoland_historical_points_interpolation}
\title{Historical points (coordinates) interpolation}
\usage{
meteoland_historical_points_interpolation(object, sf, dates, points_id)
}
\arguments{
\item{object}{\code{lfcMeteoland} object, as created by
\code{\link{meteoland}}}

\item{sf}{sf object with the the point features to interpolate.}

\item{dates}{character vector of length 2 with the dates range (start-end) to
interpolate (i.e. \code{c("2020-04-25", "2020-04-30")}). See details for
more information.}
}
\value{
An sf object.
}
\description{
\code{meteoland_points_interpolation} is a wrapper for the
  \code{$points_interpolation} method of \code{lfcMeteoland} objects.
  See also \code{\link{meteoland}}.
}
\details{
Dates must be provided as a two elements character vector, with
 the start date and the end date in a format accepted by
 \code{\link[base]{as.Date}}.
 The allowed range for dates is one natural year (365 days) ending on the
 actual date minus one day.
 Interpolation for points is made based on a 30x30 meters topology grid.#'
}
\examples{
if (interactive()) {
  library(lfcdata)
  meteolanddb <- meteoland()
  sf_points <- nfi()$get_data('plots', spatial = TRUE) \%>\%
  dplyr::slice(1:5) \%>\%
  dplyr::select(plot_id)

  meteoland_historical_points_interpolation(
    meteolanddb, sf_points, c(Sys.Date()-1, Sys.Date()-2, "plot_id")
  )
}


}
\seealso{
Other Meteoland functions: 
\code{\link{meteoland_get_lowres_raster}()},
\code{\link{meteoland_points_interpolation}()},
\code{\link{meteoland_raster_interpolation}()}
}
\concept{Meteoland functions}
