% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcMeteoland.R
\name{meteoland_raster_interpolation}
\alias{meteoland_raster_interpolation}
\title{Current raster interpolation}
\usage{
meteoland_raster_interpolation(object, sf, dates)
}
\arguments{
\item{object}{\code{lfcMeteoland} object, as created by
\code{\link{meteoland}}}

\item{sf}{sf object with the the polygon/multipolygons features to
interpolate.}

\item{dates}{character vector of length 2 with the dates range (start-end) to
interpolate (i.e. \code{c("2020-04-25", "2020-04-30")}). See details for
more information.}
}
\value{
A list of raster objects (\code{\link[raster]{raster}}), each date
  as an element of that list.
}
\description{
\code{meteoland_raster_interpolation} is a wrapper for the
  \code{$raster_interpolation} method of \code{lfcMeteoland} objects.
  See also \code{\link{meteoland}}.
}
\details{
Dates must be provided as a two elements character vector, with
 the start date and the end date in a format accepted by
 \code{\link[base]{as.Date}}.
 The allowed range for dates is one natural year (365 days) ending on the
 actual date minus one day.
 Interpolation for polygons is made based on a 1000x1000 meters topology
 raster.
}
\seealso{
Other Meteoland functions: 
\code{\link{meteoland_get_lowres_raster}()},
\code{\link{meteoland_historical_points_interpolation}()},
\code{\link{meteoland_points_interpolation}()}
}
\concept{Meteoland functions}
