% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcNFI.R
\name{nfi}
\alias{nfi}
\title{lfcNFI class}
\usage{
nfi()
}
\value{
An \code{lfcNFI} class object (inherits from
  \code{\link[R6]{R6Class}}), with methods to access the data. See Methods
  section.
}
\description{
\code{nfi()} creates an object to access the nfi database.
}
\section{Methods}{

  \code{lfcNFI} objects has the following public methods:
  \itemize{
    \item{\code{$get_data}: Retrieve and collect NFI database tables. See
          \code{\link{nfi_get_data}} for more details}
    \item{\code{$avail_tables}: Return a character vector with the names of
          the available tables in the database. See
          \code{\link{nfi_avail_tables}} for more details}
    \item{\code{$describe_table}: Print the information available about the
          provided table. See \code{\link{nfi_describe_table}} for more
          details}
    \item{\code{$describe_var}: Print the information available about the
          provided variable. See \code{\link{nfi_describe_var}} for more
          details}
  }
}

\examples{
nfidb <- nfi()
nfidb
}
\seealso{
Other NFI functions: 
\code{\link{nfi_avail_tables}()},
\code{\link{nfi_describe_table}()},
\code{\link{nfi_describe_var}()},
\code{\link{nfi_get_data}()}
}
\concept{NFI functions}
