% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcNFI.R
\name{nfi_avail_tables}
\alias{nfi_avail_tables}
\title{Get the available tables in NFI db}
\usage{
nfi_avail_tables(object)
}
\arguments{
\item{object}{\code{lfcNFI} object, as created by \code{\link{nfi}}}
}
\value{
A character vector with the table names
}
\description{
\code{nfi_avail_tables} is a wrapper for the \code{$avail_tables}
  method of \code{lfcNFI} objects. See \code{\link{nfi}}.
}
\examples{
if (interactive()) {
  nfidb <- nfi()
  nfi_avail_tables(nfidb)

  # nfidb is an R6 object, so the previous example is the same as:
  nfidb$avail_tables()
}

}
\seealso{
Other NFI functions: 
\code{\link{nfi_describe_table}()},
\code{\link{nfi_describe_var}()},
\code{\link{nfi_get_data}()},
\code{\link{nfi}()}
}
\concept{NFI functions}
