% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcNFI.R
\name{nfi_describe_table}
\alias{nfi_describe_table}
\title{Print info about the tables present in the NFI db}
\usage{
nfi_describe_table(object, tables)
}
\arguments{
\item{object}{\code{lfcNFI} object, as created by \code{\link{nfi}}}

\item{tables}{character vector with the names of the tables to describe}
}
\value{
Description is printed in the console, nothing is returned
}
\description{
\code{nfi_describe_table} is a wrapper for the \code{$describe_table}
  method of \code{lfcNFI} objects. See \code{\link{nfi}}.
}
\examples{
if (interactive()) {
  nfidb <- nfi()
  nfi_describe_table(nfidb, "plots")
  nfi_describe_table(nfidb, c("plots", "plot_nfi_4_results"))

  # nfidb is an R6 object, so the previous example is the same as:
  nfidb$describe_table("plots")
  nfidb$describe_table(c("plots", "plot_nfi_4_results"))
}

}
\seealso{
Other NFI functions: 
\code{\link{nfi_avail_tables}()},
\code{\link{nfi_describe_var}()},
\code{\link{nfi_get_data}()},
\code{\link{nfi}()}
}
\concept{NFI functions}
