% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcNFI.R
\name{nfi_describe_var}
\alias{nfi_describe_var}
\title{Print info about the variables present in the NFI db}
\usage{
nfi_describe_var(object, variables)
}
\arguments{
\item{object}{\code{lfcNFI} object, as created by \code{\link{nfi}}}

\item{variables}{character vector with the names of the variables to describe}
}
\value{
Description is printed in the console, nothing is returned
}
\description{
\code{nfi_describe_var} is a wrapper for the \code{$describe_var}
  method of \code{lfcNFI} objects. See \code{\link{nfi}}.
}
\examples{
if (interactive()) {
  nfidb <- nfi()
  nfi_describe_var(nfidb, "density")
  nfi_describe_var(nfidb, c("over_bark_volume", "basal_area"))

  # nfidb is an R6 object, so the previous example is the same as:
  nfidb$describe_var("density")
  nfidb$describe_var(c("over_bark_volume", "basal_area"))
}

}
\seealso{
Other NFI functions: 
\code{\link{nfi_avail_tables}()},
\code{\link{nfi_describe_table}()},
\code{\link{nfi_get_data}()},
\code{\link{nfi}()}
}
\concept{NFI functions}
