% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{\%>\%}
\alias{\%>\%}
\title{Funtions imported from other packages}
\description{
Funtions imported from other packages
}
\section{Pipe}{

  Exporting pipe operator from \code{\link[magrittr]{pipe}} in the magrittr package
}

