% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcSiteDrought.R
\name{siteDrought}
\alias{siteDrought}
\title{lfcsiteDrought Class}
\usage{
siteDrought()
}
\value{
An \code{siteDrought} class object (inherits from
   \code{\link[R6]{R6Class}}) with methods to access the data.
   See Methods section.
}
\description{
\code{siteDrought()} creates and object to access the siteDrought database
}
\section{Methods}{

   \code{lfcsiteDrought} objects has the following public methods.
   \itemize{
       \item {\code{$get_data} Retrieve and siteDrought database tables.
              See \code{\link{siteDrought_get_data}} for more details.}

       \item {\code{$avail_tables} bla,bla,bla}

       \item {\code{$describe_table} bla,bla,bla}

       \item {\code{$describe_var} bla,bla,bla}

       \item {\code{$print} bla,bla,bla}

   }
}

