% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcSiteDrought.R
\name{siteDrought_describe_table}
\alias{siteDrought_describe_table}
\title{Print info abuout the tables present in the siteDrought database}
\usage{
siteDrought_describe_table(object, tables)
}
\arguments{
\item{object}{\code{lfcsiteDrought} object, as created by \code{\link{siteDrought}}}

\item{tables}{character vector with the names of the tables to describe}
}
\value{
Description is printed in the console, nothing is returned
}
\description{
\code{siteDrought_describe_table} is a wrapper for the \code{$describe_table} method of
   \code{lfcsiteDrought} objects. See also \code{\link{siteDrought}}
}
\examples{

   siteDroughtdb <- lfcdata::siteDrought()
   lfcdata::siteDrought_describe_table(siteDroughtdb , 'data_day')

   # siteDrought is an R6 object, so the previus example is the same as:
   siteDroughtdb$describe_table('data_day')


}
\seealso{

}
