% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcSiteDrought.R
\name{siteDrought_describe_var}
\alias{siteDrought_describe_var}
\title{Pint info about the variables present in the siteDrought databases}
\usage{
siteDrought_describe_var(object, variables)
}
\description{
\code{siteDrought_describe_var} is a wrapper for the \code{$describe_var} method of
   \code{lfcsiteDrought} objects. See also \code{\link{siteDrought}}.
}
