% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcSiteDrought.R
\name{siteDrought_get_data}
\alias{siteDrought_get_data}
\title{Access to the tables in the siteDrought database}
\usage{
siteDrought_get_data (object, table_name)
}
\arguments{
\item{object}{\code{lfcsiteDrought} object, as created by \code{\link{siteDrought}}}

\item{table_name}{character vector of lenght 1 indicating the requested table name}
}
\value{
A tbl object, sf type
}
\description{
\code{siteDrought_get_data} is a wrapper for the \code{$get_data} method
   of \code{lfcsiteDrought} objects. See also \code{\link{siteDrought}}.
}
\details{
Connection to database can be slow. Tables retrieved from the db are stored in a temporary cache inside
   the lfcsiteDrought object created by \code{\link{siteDrought}}
}
\examples{
if (interactive()) {

  siteDroughtdb <- lfcdata::siteDrought()

  # sf tibble
  lfcdata::siteDrought_get_data(siteDroughtdb,'data_day')
  # we can uses pipes
  siteDroughtdb \%>\%
     lfcdata::siteDrought_get_data('data_day')

  # siteDroughtdb is an R6 object, so the previous examples are the same as:
  siteDroughtdb$get_data('data_day')

}
}
\seealso{
Other siteDrought functions:
   \code{\link{siteDrought_avail_tables}} (),
   \code{\link{siteDrought_describe_table}} (),
   \code{\link{siteDrought_describe_var}} (),
   \code{\link{siteDrought}} ()
}
