% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcSiteDrought.R
\name{str_unique}
\alias{str_unique}
\title{Detect the a pattern in the same way as `SQL`'s `LIKE` operator.}
\usage{
str_unique(string, ...)
}
\arguments{
\item{string}{A character vector to return unique entries.}

\item{...}{Other options used to control matching behaviour between
duplicate strings. Passed on to [stringi::stri_opts_collator()].}
}
\value{
A character vector.
}
\description{
`str_like()` follows the conventions of the SQL `LIKE` operator:

* Must match the entire string.
* `_` matches a single character (like `.`).
* `%` matches any number of characters (like `.*`).
* `\%` and `\_` match literal `%` and `_`.
* The match is case insensitive by default.
}
\examples{
str_unique(c("a","b","c","a","a","f"))

# Use ... to pass additonal arguments to stri_unique()
str_unique(c("motley", "mötley", "pinguino", "pingüino"))
str_unique(c("motley", "mötley", "pinguino", "pingüino"), strength = 1)
}
\seealso{
[unique()], [stringi::stri_unique()] witch this funcions wraps.
}
