% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcSiteDrought.R
\name{str_unique}
\alias{str_unique}
\title{Remove duplicates strings}
\usage{
str_unique(string, ...)
}
\arguments{
\item{string}{A character vector to return unique entries.}

\item{...}{Other options used to control matching behaviour between
duplicate strings. Passed on to [stringi::stri_opts_collator()].}
}
\value{
A character vector.
}
\description{
`str_unique()` remove duplicate values, with optional control over
how duplicte is measured.
}
\details{
`str_detect()` return a logial vector `TRUE` if `pattern` or `FALSE` if bla bla bla
}
\examples{
str_unique(c("a","b","c","a","a","f"))

# Use ... to pass additonal arguments to stri_unique()
str_unique(c("motley", "mötley", "pinguino", "pingüino"))
str_unique(c("motley", "mötley", "pinguino", "pingüino"), strength = 1)
}
\seealso{
[unique()], [stringi::stri_unique()] witch this funcions wraps.
}
